fix: Handle Github Oauth signup. Fixes #162
Our Oath signup flow for Github assumed the user's email address
would be returned in the API call fetching the user info.
As it happens, that only returns the email address if/when the
user has listed their address as *public* on the Github site.

For people who haven't, and extra API call is required to fetch
the user's email address(es) as well as an extra Oauth scope.

This implements that on the backend, the extra scope is configured
in the frontend.